+++ main.go
@@ -107,196 +107,3 @@ func main() {
 	// Use the formatted address to listen
 	log.Fatal(http.ListenAndServe(serverAddr, nil))
 }
-
-func getSchoolsHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	// LATER: connect to database, extract corresponding list and parse it
-	schools := getSchools()
-
-	// Convert to Json
-	if err := json.NewEncoder(w).Encode(schools); err != nil {
-		JSONError(w, "Failed to encode schools response", http.StatusInternalServerError)
-		log.Printf("Error encoding response: %v", err)
-		return
-	}
-	log.Printf("Successfully served /api/schools request")
-}
-
-func getGradesHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	// Extract the school_id query parameter
-	schoolID := r.URL.Query().Get("school_id")
-
-	// 1. Input Validation: Check if the required parameter is missing
-	if schoolID == "" {
-		JSONError(w, "Missing required query parameter: school_id", http.StatusBadRequest)
-		return
-	}
-
-	log.Printf("Received request for grades in school ID: %s", schoolID)
-
-	// LATER: The mock data here would be filtered based on schoolID
-	// For now, we return the full mock list regardless of the ID.
-
-	// LATER: connect to database, extract corresponding list and parse it
-
-	grades := getGrades(schoolID)
-
-	// Convert to Json
-	if err := json.NewEncoder(w).Encode(grades); err != nil {
-		JSONError(w, "Failed to encode grades response", http.StatusInternalServerError)
-		log.Printf("Error encoding response: %v", err)
-		return
-	}
-	log.Printf("Successfully served /api/grades request")
-}
-
-func getEquipmentListsHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	// Extract the required query parameters (updated)
-	schoolID := r.URL.Query().Get("school_id")
-	gradeID := r.URL.Query().Get("grade_id")
-
-	// 1. Input Validation (updated)
-	if schoolID == "" || gradeID == "" {
-		JSONError(w, "Missing required query parameters: school_id or grade_id", http.StatusBadRequest)
-		return
-	}
-
-	log.Printf("Received request for equipment list: School=%s, Grade=%s", schoolID, gradeID)
-
-	// LATER: connect to database, extract corresponding list and parse it
-	equipment := getEquipment(schoolID, gradeID)
-
-	response := EquipmentListResponse{
-		Items: equipment,
-	}
-
-	if err := json.NewEncoder(w).Encode(response); err != nil {
-		JSONError(w, "Failed to encode equipment response", http.StatusInternalServerError)
-		log.Printf("Error encoding response: %v", err)
-		return
-	}
-	log.Printf("Successfully served /api/equipment request")
-}
-
-// =====NEW=====
-// adding handlers to login page & shopping cart
-func authStatusHandler(w http.ResponseWriter, r *http.Request) {
-	cookie, err := r.Cookie("sessionid")
-	if err != nil {
-		JSONError(w, "Unauthorized", http.StatusUnauthorized)
-		return
-	}
-
-	userID, exists := sessions[cookie.Value]
-	if !exists {
-		JSONError(w, "Unauthorized", http.StatusUnauthorized)
-		return
-	}
-
-	var username = getUsernameFromUserID(userID)
-	if username != "" {
-		err := json.NewEncoder(w).Encode(map[string]string{"userid": userID, "username": username})
-		if err != nil {
-			log.Printf("Failed to encode auth status response: %v", err)
-		}
-		return
-	} else {
-		JSONError(w, "Unauthorized", http.StatusUnauthorized)
-	}
-}
-
-func postLoginHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		JSONError(w, "Method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	var credentials struct {
-		Username string `json:"username"`
-		Password string `json:"password"`
-	}
-
-	if err := json.NewDecoder(r.Body).Decode(&credentials); err != nil {
-		JSONError(w, "Failed to decode request body", http.StatusBadRequest)
-		return
-	}
-
-	var userId = getUserIDByCredentials(credentials.Username, credentials.Password)
-
-	if userId == "" {
-		JSONError(w, "Incorrect username or password. Please try again.", http.StatusUnauthorized)
-	} else {
-		// Session generation
-		sessionID := generateSessionID()
-		sessions[sessionID] = userId
-
-		// Cookie setting
-		http.SetCookie(w, &http.Cookie{
-			Name:     "sessionid",
-			Value:    sessionID,
-			Path:     "/",
-			HttpOnly: true,
-			//Secure: true, // Uncomment this line if using HTTPS
-			//SameSite: http.SameSiteStrictMode,
-		})
-		err := json.NewEncoder(w).Encode(map[string]string{"userid": userId, "username": credentials.Username})
-		if err != nil {
-			log.Printf("Failed to encode login response: %v", err)
-		}
-		return
-	}
-}
-
-func logoutHandler(w http.ResponseWriter, r *http.Request) {
-	cookie, err := r.Cookie("sessionid")
-	if err == nil {
-		delete(sessions, cookie.Value)
-	}
-	http.SetCookie(w, &http.Cookie{
-		Name:     "sessionid",
-		Value:    "",
-		Path:     "/",
-		MaxAge:   -1,
-		HttpOnly: true,
-	})
-	w.WriteHeader(http.StatusOK)
-}
-
-func getPostCartHandler(w http.ResponseWriter, r *http.Request) {
-	userID := r.URL.Query().Get("userid")
-	if userID == "" {
-		JSONError(w, "Missing required query parameter: userid", http.StatusBadRequest)
-		return
-	}
-
-	switch r.Method {
-	case http.MethodGet:
-		// Return existing cart (now returns []CartEntry)
-		cart := getCartByUserID(userID)
-		err := json.NewEncoder(w).Encode(cart)
-		if err != nil {
-			log.Printf("Failed to encode cart response: %v", err)
-		}
-
-	case http.MethodPost, http.MethodPut:
-		// Update the cart (expects []CartEntry)
-		var newEntries []CartEntry
-		if err := json.NewDecoder(r.Body).Decode(&newEntries); err != nil {
-			JSONError(w, "Failed to decode request body", http.StatusBadRequest)
-			return
-		}
-		saveCart(userID, newEntries)
-		w.WriteHeader(http.StatusOK)
-		if _, err := fmt.Fprintf(w, "Cart updated successfully"); err != nil {
-			log.Printf("Failed to write cart update response: %v", err)
-		}
-
-	default:
-		JSONError(w, "Method not allowed", http.StatusMethodNotAllowed)
-	}
-}
